fixing errors, commenting out nonexistent features
Made some changes to fix sign readability, but lever interaction remains broken. From investigating the existing test games, it just seems that item interaction as described does not exist in the current WDL, as many of the behaviors I assumed existed were simply commented out in the existing test game. As such, this game is not ideally functional, but follows the pattern of existing test games and highlights many areas of improvement for our WDL++ changes.