Cleaned up the top-level games/ directory
This directory should only contain WDL files that can be opened by
the current version of chiventure. Any documentation on the game
should appear at the top of the WDL file.

Note: The deleted test_game_readme.txt should probably be converted
to backlog issues, since it's not really documentation of the game
but rather a list of issues encountered when writing the game.