After talking on Team Slack, decided to not include a way to offer two types of copies for stat and effect. Before we had:   - copy_stat_and_global and copy_stat         - The only difference was that one created a deep copy of both           the stat structure as well as the global at the same time while           the other only of the stat structure
Now will only have: copy_stat which only deep copies the stat structure

This was decided because it would be far too confusing for Users as they would
require external knowledgge of how the actual stat and effect structures are
featured. Also, if there is a need in the future, this can be easily implemented
as copy_stat_and_global is essentially a wrapper for copy_global_stat and copy_stat

Note: This change is also done for the copies of effect. The functions are very
similar so the explanation carries over. The only difference is that we would be
talking about effect structures rather than stat structures